Fixed the template skill to be the trivia skill
The fact skill wasn't working, in the sense that the nodejs code to be pasted in the tutorial didn't work with it. There were missing dependencies, the "questions.js" source module was missing, etc.
I am sure you intended the trivia skill here, since this is a trivia skill tutorial, and otherwise it wouldn't work at all.